pyproject.toml
[project]
name = "cfo-liquidity-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "databricks-sdk>=0.78.0",
    "functions-framework>=3.10.0",
    "google-cloud-secret-manager>=2.26.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "stripe>=14.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
